removed handling for line chart and removed unnecessary log lines
